requirements.txt
algoliasearch==4.21.0
python-dotenv
supabase
openai
streamlit
streamlit-js-eval
